Chore: Update pre-commit dependencies to latest
ERROR! the role 'protobuf-install' was not found in <>
The error appears to be in '': line 5, column 7, but may
be elsewhere in the file depending on the exact syntax
problem.

Update role name to match the role path on GHA and
Jenkins jobs.

Add GHA verify workflow for ansible roles.
Update tox and pre-commit dependencies.

Fix "Error: Unable to find a match: python2-dnf".

Ref: https://github.com/ansible/molecule/issues/1042
Change-Id: Iff5078c54ddc5768b6d2540129df17275c83d77b
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>